re-arrange unit tests
